requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2

